+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 1.0.0a15 (unreleased)
 ---------------------
 
+- Remove get_current_request optimizations until memory leaks can be solved
+  [vangheem]
+
 - Provide more robust conflict resolution on fields of content and annotations
   [vangheem]